+++ lang/en/theme_foundation.php
@@ -73,6 +73,12 @@ $string['generalheadingdesc'] = 'Configure the general settings for Foundation h
 $string['navbarposition'] = 'Navbar position';
 $string['navbarpositiondesc'] = 'Set the position of the navbar.';
 
+$string['footerantigravityhorizontaloffset'] = 'Screen anti-gravity horizontal offset';
+$string['footerantigravityhorizontaloffsetdesc'] = 'Set the horizontal offset of the anti-gravity buttons on the screen (not the navbar) from the default.  This can be an positive (move right) or negative (move left) number.  The units are pixels.  Note:  Flip this logic for RTL languages.';
+
+$string['footerantigravityverticaloffset'] = 'Screen anti-gravity vertical offset';
+$string['footerantigravityverticaloffsetdesc'] = 'Set the vertical offset of the anti-gravity buttons on the screen (not the navbar) from the default.  This can be an positive (move down) or negative (move up) number.  The units are pixels.';
+
 $string['favicon'] = 'Custom favicon';
 $string['favicondesc'] = 'Upload your own favicon.  It should be an .ico file.';